requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cryptography==45.0.6
decorator==5.2.1
exceptiongroup==1.3.0
filelock==3.13.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==2.1.5
mpmath==1.3.0
narwhals==2.1.2
networkx==3.2.1
numpy==1.24.3
packaging==25.0
pandas==1.5.3
Pillow==9.5.0
protobuf==6.32.0
py==1.11.0
pyarrow==21.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
retry==0.9.2
rpds-py==0.27.0
scikit-learn==1.2.2
scipy==1.13.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
streamlit==1.48.1
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
volcengine==1.0.196
volcengine-python-sdk==4.0.11
watchdog==6.0.0
